+++ SDeMo/src/calibration/isotonic.jl
@@ -93,3 +93,11 @@ function PAVA(x, y, w=ones(length(y)))
 end
 
 correct(is::IsotonicCalibration, y) = is.calibrator(y)
+
+@testitem "We can do Isotonic calibration" begin
+    X, y = SDeMo.__demodata()
+    model = SDM(PCATransform, NaiveBayes, X, y)
+    train!(model)
+    C = calibrate(IsotonicCalibration, model; bins=12)
+    @test typeof(C) <: IsotonicCalibration
+end